Installed react-number-format to use in amountInput
