+++ package.json
@@ -34,6 +34,7 @@
     "react-dom": "^18.1.0",
     "react-icons": "^5.5.0",
     "react-modal-sheet": "^4.0.1",
+    "react-number-format": "^5.4.4",
     "react-router": "^7.3.0",
     "tauri-plugin-androidpermissionprompt-api": "^0.1.1",
     "tinybase": "^5.4.9",
